LPS-60972 "directDeploy" tasks can use the "portal" configuration
