+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/tasks/DirectDeployTask.java
@@ -14,6 +14,7 @@
 
 package com.liferay.gradle.plugins.tasks;
 
+import com.liferay.gradle.plugins.LiferayJavaPlugin;
 import com.liferay.gradle.plugins.util.FileUtil;
 import com.liferay.gradle.util.GradleUtil;
 import com.liferay.gradle.util.StringUtil;
@@ -24,10 +25,12 @@ import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.gradle.api.artifacts.Configuration;
+
 /**
  * @author Andrea Di Giorgi
  */
-public class DirectDeployTask extends BasePortalImplToolsTask {
+public class DirectDeployTask extends BasePortalToolsTask {
 
 	public File getAppServerDeployDir() {
 		return GradleUtil.toFile(project, _appServerDeployDir);
@@ -35,6 +38,14 @@ public class DirectDeployTask extends BasePortalImplToolsTask {
 
 	public File getAppServerDir() {
 		return GradleUtil.toFile(project, _appServerDir);
+	}
+
+	public File getAppServerLibGlobalDir() {
+		return GradleUtil.toFile(project, _appServerLibGlobalDir);
+	}
+
+	public File getAppServerPortalDir() {
+		return GradleUtil.toFile(project, _appServerPortalDir);
 	}
 
 	public String getAppServerType() {
@@ -167,6 +178,14 @@ public class DirectDeployTask extends BasePortalImplToolsTask {
 		_appServerDir = appServerDir;
 	}
 
+	public void setAppServerLibGlobalDir(Object appServerLibGlobalDir) {
+		_appServerLibGlobalDir = appServerLibGlobalDir;
+	}
+
+	public void setAppServerPortalDir(Object appServerPortalDir) {
+		_appServerPortalDir = appServerPortalDir;
+	}
+
 	public void setAppServerType(Object appServerType) {
 		_appServerType = appServerType;
 	}
@@ -193,7 +212,13 @@ public class DirectDeployTask extends BasePortalImplToolsTask {
 
 	@Override
 	protected void addDependencies() {
-		super.addDependencies();
+		Configuration configuration = GradleUtil.getConfiguration(
+			project, getConfigurationName());
+
+		Configuration portalConfiguration = GradleUtil.getConfiguration(
+			project, LiferayJavaPlugin.PORTAL_CONFIGURATION_NAME);
+
+		configuration.extendsFrom(portalConfiguration);
 
 		String appServerType = getAppServerType();
 
@@ -201,8 +226,13 @@ public class DirectDeployTask extends BasePortalImplToolsTask {
 			File dir = new File(getAppServerDir(), "lib/endorsed");
 
 			GradleUtil.addDependency(
-				project, getConfigurationName(), getJarsFileTree(dir));
+				project, configuration.getName(),
+				FileUtil.getJarsFileTree(project, dir));
 		}
+	}
+
+	protected File getAppServerLibPortalDir() {
+		return new File(getAppServerPortalDir(), "WEB-INF/lib");
 	}
 
 	@Override
@@ -212,6 +242,8 @@ public class DirectDeployTask extends BasePortalImplToolsTask {
 
 	private Object _appServerDeployDir;
 	private Object _appServerDir;
+	private Object _appServerLibGlobalDir;
+	private Object _appServerPortalDir;
 	private Object _appServerType;
 	private Object _argAppServerType;
 	private boolean _customPortletXml;
